Critical and public response

Format

A data frame with 31 rows and 17 columns:

film

name of film

rotten_tomatoes_score

score from the American review-aggregation website Rotten Tomatoes; scored out of 100

rotten_tomatoes_counts

number of critics contributing to Rotten Tomatoes score

metacritic_score

score from Metacritic where scores are weighted average of reviews; scored out of 100

metacritic_counts

number of critics contributing to Metacritic score

cinema_score

score from market research firm CinemaScore; scored by grades A, B, C, D, and F

imdb_score

score from IMDb where scores are weighted average of reviews; scored out of 100

imdb_counts

number of critics contributing to IMDb score

rotten_tomatoes_score_top_critics

score for only top critics from Rotten Tomatoes; scored out of 100

rotten_tomatoes_score_top_critics_votes

number of top critics contributing to Rotten Tomatoes score

rotten_tomatoes_popcorn_meter_score

percentage of users who rated rotten_tomatoes_popcorn_meter_votesnumber of users contributing to Rotten Tomatoes' rotten_tomatoes_popcorn_meter_verifiedpercentage of users who made a verified movie ticket purchase rating this 3.5 stars or higher rotten_tomatoes_popcorn_meter_votesnumber of users contributing to Rotten Tomatoes' Popcorn Meter letterboxd_ratingscore from Letterboxd where scores are weighted average of reviews; scored out of 5 letterboxd_countsnumber of users contributing to Letterboxd score letterboxd_fansnumber of users this film as one of their four favorites on Letterboxd

https://en.wikipedia.org/wiki/List_of_Pixar_films public_response A data set with scores of critical and public response. public_response datasets